Starting to add DiracSpinor: working, but program does double work. Starting full transition now

+++ src/AKF_akFunctions.h
@@ -19,14 +19,14 @@ int akReadWrite(std::string fname, bool write,
                 double &dEmin, double &dEmax);
 
 int calculateK_nk(const ElectronOrbitals &wf, int nk, int max_L, double dE,
-                  std::vector<std::vector<std::vector<float>>> &jLqr_f,
+                  std::vector<std::vector<std::vector<double>>> &jLqr_f,
                   std::vector<float> &K_nk, double Zeff = -1);
 
 int calculateKpw_nk(const ElectronOrbitals &wf, int nk, double dE,
-                    std::vector<std::vector<float>> &jl_qr,
+                    std::vector<std::vector<double>> &jl_qr,
                     std::vector<float> &K_nk);
 
-void sphericalBesselTable(std::vector<std::vector<std::vector<float>>> &jLqr_f,
+void sphericalBesselTable(std::vector<std::vector<std::vector<double>>> &jLqr_f,
                           int max_L, const std::vector<double> &q_array,
                           const std::vector<double> &r);
